Add option to selectively enable animated heights
Animating an accordion's height as it's opening/closing can appear jolty
on mobile devices, especially if it's pushing down a lot of content. The
animHeight option was added to deal with this, which sets an accordion's
height without tweening. However, the tradeoff is that users can see the
content under an accordion "jumping" as it's opening/closing. Unless the
widget is situated somewhere where there's no following content, this is
a serious detriment to UX.

The "animHeight" option can now be assigned the special value of "auto",
which sets animated heights based on whether surrounding content will be
visible after a fold's finished opening (or closing).