Delay setting an open fold's height on page load
If a fold is marked as open on page load, and the accordion it's sitting
inside is nested within another, heights are measured incorrectly. While
the reason for this isn't 100% clear, it might have something to do with
every accordion's resize method firing at the same time.

Setting a pre-opened fold's height after a brief delay seems to fix this
issue, irrespective of what's causing it. Delays less than 50 ms seem to
work sporadically on Firefox and IE, which gives reason to believe there
are far better solutions to this issue.

Tentatively fixes gh-3.