Revert "Add handlers to update folds when images load"
Unsatisfied with the solution attempted by 6c13231, since I feel this is
something best left to a developer's discretion. Not every accordion has
replaced elements inside it, and not every replaced element will need to
have their dimensions evaluated (such as if CSS is used to set the width
and height of the image's container).