Fix bug with grandparent accordions not resizing
When an accordion is instantiated, it calls its parent's "update" method
to adjust the height of the containing fold. Doing this will also update
the heights of any additional ancestors which contain the parent.

However, a grandchild accordion was also attempting to walk the ancestry
and update each ancestor individually, which conflicted with the updates
that an accordion's "update" method already takes care of.