Add options to configure behaviour of accordions
This commit restores the ability to adjust which CSS classes are applied
to DOM elements, and disable ancillary functionality like ARIA attribute
management.

Fold instances now read options from the accordions they're attached to,
instead of accepting a hash of options individually. Because they're not
expected to be created by or interfaced with external code, it's useless
to allow this level of configurability.