Transpile classes to ECMAScript 5
Because of the bizarre hack used to get Object.defineProperty working on
IE8, we can't use Babel to transpile ES6 to ES5. The function which gets
things working on IE8 returns a copy of the original constructor object,
which has to replace the original in the current scope. For this to work
in Babel would require lots of weird/ugly hacking with its extensions.

Sadly, this also means that every future change made to the original ES6
files will need to be applied to the transpiled ES5 file, too.