Check scroll offsets have changed when tapping folds
Safari on iOS doesn't set a TouchEvent's "cancelable" property to false.
This means iPhone users were still subjected to the issue 8d498bc sought
to address. We can work around this by checking if the viewport's scroll
offsets have changed between the gesture's start and end events. If they
have, we should interpret that as a user's finger scrolling the content.