Add option to check borders when calculating height
Calculating border widths is computationally more expensive than using a
fixed, predetermined value. If the size of a fold's borders are constant
and known in advance, consider using the "heightOffset" setting instead.
Remember to set "useBorders" to false if doing so.

If set to "auto" (the default), the value is set based on whether or not
the fold's elements appear to have a border when it's being created. For
performance's sake, it's always best to explicitly disable the option if
it's known that borders aren't being used.