Ditch the idea of "different module flavours"
The whole `dist` directory shouldn't have been added in the first place.
It's basically spoon-feeding authors a bunch of copy-and-pasted that are
already available from the `src` directory.

AMD support is axed altogether in favour of full ESM support. CommonJS's
support has been moved to the ES5 version of the file (src/accordion.js)
where it extends the `module.exports` object, if present.

These removals are breaking changes, so the next release will be a major
version bump.