Add ESLint support for `.mjs` files
I'm refraining from running lint on the original ES5 source, because the
syntax purposefully uses irregularities for improved munging when uglify
compresses it (this infamous habit of mine died somewhere in late-2015).

Echoes of my spastic JS acrobatics are still visible in helpers.mjs, but
I haven't the heart to change code with a nostalgic tint in my memory.